Improve min_width calculation using longest word
Instead of using the styled length of 'M', use the styled width of the longest word in order to calculate the min_width of a Cell's Text.

This will prevent words from being split in the middle.

There may be an issue with this if the total of the min_widths add up to more than the available width. We should look into the tests around that to see if something will error or if this is being accounted for already.