FIX: Handle the case where cell content has inline formatting.
This patch changes the way the longest word in the table cell is calculated
to account for table cell that contain 'text' with inline formatting.

This is done by using Prawn's functionality to parse a string into it's
formatted parts and then collecting all the text without the formatting
information and selecting the longest word from that collection.